fix: GREEN - resolve status.go:122 change ToolBudgetBytes to *int64 for consistent byte-count types

+++ internal/registry/config_test.go
@@ -357,7 +357,7 @@ agents:
 	require.NotNil(t, a.MaxTurns)
 	assert.Equal(t, 5, *a.MaxTurns, "max_turns parsed")
 	require.NotNil(t, a.ToolBudgetBytes)
-	assert.Equal(t, 4096, *a.ToolBudgetBytes, "tool_budget_bytes parsed")
+	assert.Equal(t, int64(4096), *a.ToolBudgetBytes, "tool_budget_bytes parsed")
 	assert.Equal(t, "skeptic", a.Role, "role parsed")
 }
